pyproject.toml
[tool.poetry]
name = "repl_python3_LongtermJitteryCrypto"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
praw = "^7.2.0"
aiohttp = "^3.7.4"
asyncpraw = "^7.2.0"
prsaw = "^0.3.1"
flask = "^1.1.2"
"statcord.py" = "^3.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
